requirements.txt
confluent-kafka>=2.8.0
python-dotenv>=1.0.1
